refactor(components/imports): add proper import paths for layouts
because of https://github.com/Darosss/TwitchChatBot/commit/cdb0ffefb101329e216bf5796197d97b2402dfd7 - where index.ts files are created - change import paths to src/layout
